# firebase-react/functions/src/routes/v1/approvals.ts
import type express from "express";
import type * as admin from "firebase-admin";
import crypto from "node:crypto";

import { requireAuth, requireApprover, requireOps, roleOf } from "../../lib/auth";
import { fail, ok } from "../../lib/http";
import { withIdempotency } from "../../lib/idempotency";
import { approvalRef, decideApproval } from "../../lib/approvals";
import { writeTimelineEvent } from "../../lib/timeline";

function quoteRef(adminApp: typeof admin, caseId: string, quoteId: string) {
  return adminApp.firestore().doc(`cases/${caseId}/quotes/${quoteId}`);
}

export function registerApprovalRoutes(app: express.Express, adminApp: typeof admin) {
  // ops: approvals list
  app.get("/v1/ops/approvals", async (req, res) => {
    const auth = await requireAuth(adminApp, req, res);
    if (!auth) return;
    if (!requireOps(auth)) return fail(res, 403, "FORBIDDEN", "운영자 권한이 없습니다.");

    const status = String(req.query.status || "pending");
    const gate = req.query.gate ? String(req.query.gate) : null;
    let q = adminApp.firestore().collection("approvals").where("status", "==", status);
    if (gate) q = q.where("gate", "==", gate);
    q = q.orderBy("createdAt", "asc").limit(50);

    const snap = await q.get();
    const items = snap.docs.map((d) => ({ id: d.id, ...d.data() }));
    return ok(res, { items });
  });

  // ops_approver: decision + post-actions
  app.post("/v1/ops/approvals/:approvalId/decision", async (req, res) => {
    const auth = await requireAuth(adminApp, req, res);
    if (!auth) return;
    if (!requireApprover(auth)) return fail(res, 403, "FORBIDDEN", "승인 권한이 없습니다.");

    const approvalId = req.params.approvalId;
    const { decision, reasonKo } = req.body ?? {};
    if (decision !== "approve" && decision !== "reject") return fail(res, 400, "INVALID_ARGUMENT", "decision이 필요합니다.");

    const result = await withIdempotency(adminApp, req, res, "approvals.decision", async () => {
      const ref = approvalRef(adminApp, approvalId);
      const snap = await ref.get();
      if (!snap.exists) throw new Error("NOT_FOUND");
      const a = snap.data() as any;

      await decideApproval(
        adminApp,
        approvalId,
        decision === "approve" ? "approve" : "reject",
        { uid: auth.uid, role: roleOf(auth) },
        reasonKo
      );

      // 케이스 타임라인 이벤트
      const target = a.target as any;
      if (target?.caseId) {
        const now = adminApp.firestore.FieldValue.serverTimestamp();
        const eventId = crypto.randomUUID();
        await writeTimelineEvent(adminApp, target.caseId, eventId, {
          type: decision === "approve" ? "APPROVAL_APPROVED" : "APPROVAL_REJECTED",
          occurredAt: now,
          actor: { type: "ops", uid: auth.uid },
          summaryKo: decision === "approve" ? "승인되었습니다." : "반려되었습니다.",
          meta: { approvalId, gate: a.gate, target: a.target, reasonKo: reasonKo ?? null }
        });
      }

      // post-action: quote finalize 자동 반영(승인 시)
      if (decision === "approve" && a.gate === "quote_finalize") {
        const t = a.target as any;
        if (t?.caseId && t?.quoteId) {
          const now = adminApp.firestore.FieldValue.serverTimestamp();
          await quoteRef(adminApp, t.caseId, t.quoteId).set({ status: "finalized", finalizedAt: now, updatedAt: now }, { merge: true });
          const eventId = crypto.randomUUID();
          await writeTimelineEvent(adminApp, t.caseId, eventId, {
            type: "QUOTE_FINALIZED",
            occurredAt: now,
            actor: { type: "ops", uid: auth.uid },
            summaryKo: "견적이 확정되었습니다.",
            meta: { quoteId: t.quoteId, approvalId }
          });
        }
      }

      return { status: decision === "approve" ? "approved" : "rejected" };
    }).catch((e: any) => {
      if (String(e?.message) === "NOT_FOUND") {
        fail(res, 404, "NOT_FOUND", "승인 요청을 찾을 수 없습니다.");
        return null;
      }
      fail(res, 500, "INTERNAL", "서버 오류가 발생했습니다.");
      return null;
    });

    if (!result) return;
    return ok(res, result);
  });
}